feat: add JWTManager::decode()
